+++ libvirt-gobject/libvirt-gobject-network-filter.h
@@ -66,7 +66,7 @@ const gchar *gvir_network_filter_get_uuid(GVirNetworkFilter *filter);
 
 GVirConfigNetworkFilter *gvir_network_filter_get_config
                                 (GVirNetworkFilter *filter,
-                                 guint64 flags,
+                                 guint flags,
                                  GError **err);
 
 G_END_DECLS
